requirements.txt
streamlit
pinecone-client
cohere
torch
transformers
sentence-transformers
PyPDF2==3.0.1
